requirements.txt
pandas
numpy
scikit-learn
joblib
matplotlib
pyserial
pyyaml
